pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Virgin Voyages MXP MCP Server with Database Access"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi-mcp>=0.3.4",
    "pymssql>=2.3.0",
    "python-dotenv>=1.0.0",
]

[tool.basedpyright]
typeCheckingMode = "standard"
reportExplicitAny = false
reportMissingTypeStubs = false
